change code locations
