+++ worker/hostcalls/common/common.go
